# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axiom Lumen Public API",
    "version": "1.0.0",
    "description": "Persisted Stellar reconciliation snapshots. The production specification includes implemented routes only.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
...
  "paths": {
...
    "/api/v1/trustlines/{asset}": {
      "get": {
        "operationId": "getTrustlines",
        "tags": [
          "Trustlines"
        ],
        "summary": "Get finalized Public Network trustline authorization-state counts",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Asset"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Current verified or degraded trustline-state snapshot",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationSnapshot"
                },
                "examples": {
                  "verified": {
                    "$ref": "#/components/examples/TrustlineVerified"
                  },
                  "degraded": {
                    "$ref": "#/components/examples/TrustlineDegraded"
                  }
                }
              }
            }
          },
          "304": {
            "description": "The representation matches If-None-Match",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            }
          },
          "400": {
            "description": "Invalid asset, request header, or query parameter",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "invalidAsset": {
                    "$ref": "#/components/examples/InvalidAsset"
                  },
                  "invalidRequestId": {
                    "$ref": "#/components/examples/InvalidRequestId"
                  },
                  "invalidQueryParameter": {
                    "$ref": "#/components/examples/InvalidQueryParameter"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationError"
          },
          "404": {
            "description": "No finalized snapshot exists for the asset",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "missing": {
                    "$ref": "#/components/examples/TrustlineMissingSnapshot"
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          },
          "503": {
            "description": "Persisted metric state, freshness, read storage, or API access storage is unavailable",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ReconciliationSnapshot"
                    },
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ]
                },
                "examples": {
                  "unavailable": {
                    "$ref": "#/components/examples/TrustlineUnavailable"
                  },
                  "readUnavailable": {
                    "$ref": "#/components/examples/TrustlineReadUnavailable"
                  }
                }
              }
            }
          }
        }
      },
      "options": {
        "operationId": "trustlineOptions",
        "summary": "Inspect public read-only CORS policy",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "204": {
            "description": "CORS preflight accepted",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "Access-Control-Allow-Headers": {
                "$ref": "#/components/headers/CorsAllowHeaders"
              },
              "Access-Control-Allow-Methods": {
                "$ref": "#/components/headers/CorsAllowMethods"
              },
              "Access-Control-Max-Age": {
                "$ref": "#/components/headers/CorsMaxAge"
              }
            }
          },
          "400": {
            "description": "Invalid request identifier",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "Access-Control-Allow-Headers": {
                "$ref": "#/components/headers/CorsAllowHeaders"
              },
              "Access-Control-Allow-Methods": {
                "$ref": "#/components/headers/CorsAllowMethods"
              },
              "Access-Control-Max-Age": {
                "$ref": "#/components/headers/CorsMaxAge"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "invalidRequestId": {
                    "$ref": "#/components/examples/InvalidRequestId"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:-]*$"
          },
          "as_of": {
            "type": "string",
            "format": "date-time"
          },
          "api_version": {
            "type": "string",
            "const": "v1"
          }
        },
        "required": [
          "error",
          "request_id",
          "as_of",
          "api_version"
        ],
        "additionalProperties": false
      }
    },
    "responses": {
      "AuthenticationError": {
        "description": "API key is missing, malformed, expired, revoked, or inactive when authentication is required",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControl"
          },
          "Access-Control-Allow-Origin": {
            "$ref": "#/components/headers/CorsOrigin"
          },
          "Access-Control-Expose-Headers": {
            "$ref": "#/components/headers/CorsExposeHeaders"
          },
          "Vary": {
            "$ref": "#/components/headers/Vary"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "examples": {
              "authentication": {
                "$ref": "#/components/examples/AuthenticationError"
              }
            }
          }
        }
      },
      "RateLimitError": {
        "description": "The authenticated principal exhausted its plan quota for the current fixed window",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControl"
          },
          "Access-Control-Allow-Origin": {
            "$ref": "#/components/headers/CorsOrigin"
          },
          "Access-Control-Expose-Headers": {
            "$ref": "#/components/headers/CorsExposeHeaders"
          },
          "Vary": {
            "$ref": "#/components/headers/Vary"
          },
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          },
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/XRateLimitLimit"
          },
          "X-RateLimit-Remaining": {
            "$ref": "#/components/headers/XRateLimitRemaining"
          },
          "X-RateLimit-Reset": {
            "$ref": "#/components/headers/XRateLimitReset"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "examples": {
              "rateLimit": {
                "$ref": "#/components/examples/RateLimitError"
              }
            }
          }
        }
      }
    },
    "parameters": {
      "Asset": {
        "name": "asset",
        "in": "path",
        "required": true,
        "description": "Canonical classic credit asset identifier CODE:ISSUER",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9]{1,12}:G[A-Z2-7]{55}$"
        },
        "example": "USDC:GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
      },
...
      "RequestId": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "Optional caller correlation identifier; generated when absent",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:-]*$"
        }
      },
      "IfNoneMatch": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "headers": {
      "XRequestId": {
        "description": "Request correlation identifier",
        "schema": {
          "type": "string"
        }
      },
      "CacheControl": {
        "description": "Private snapshot caching or no-store",
        "schema": {
          "type": "string"
        }
      },
      "ETag": {
        "description": "Weak complete-representation validator",
        "schema": {
          "type": "string"
        }
      },
      "CorsOrigin": {
        "description": "Public read-only origin policy",
        "schema": {
          "type": "string",
          "enum": [
            "*"
          ]
        }
      },
      "CorsExposeHeaders": {
        "description": "Response headers available to browsers",
        "schema": {
          "type": "string"
        }
      },
      "CorsAllowHeaders": {
        "description": "Headers accepted by CORS requests",
        "schema": {
          "type": "string"
        }
      },
      "CorsAllowMethods": {
        "description": "Methods accepted by CORS requests",
        "schema": {
          "type": "string",
          "enum": [
            "GET, OPTIONS"
          ]
        }
      },
      "CorsMaxAge": {
        "description": "Preflight cache lifetime in seconds",
        "schema": {
          "type": "string",
          "enum": [
            "86400"
          ]
        }
      },
      "Vary": {
        "description": "Cache representation varies by request identifier and API credential",
        "schema": {
          "type": "string",
          "enum": [
            "X-Request-ID, X-Axiom-Key"
          ]
        }
      },
      "RetryAfter": {
        "description": "Seconds until a quota-limited request may be retried",
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      },
      "XRateLimitLimit": {
        "description": "Maximum requests in the current plan window",
        "schema": {
          "type": "string",
          "pattern": "^[1-9]\\d*$"
        }
      },
      "XRateLimitRemaining": {
        "description": "Requests remaining after this request",
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      },
      "XRateLimitReset": {
        "description": "UTC Unix timestamp when the fixed window resets",
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      }
    },
    "examples": {
...
      "TrustlineVerified": {
        "summary": "Verified trustline state",
        "value": {
          "metric": "trustline_state",
          "subject": {
            "kind": "asset",
            "asset": "USDC:GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
          },
          "status": "verified",
          "value": {
            "kind": "trustline_state",
            "total": "825",
            "states": {
              "authorized": "700",
              "authorized_to_maintain_liabilities": "100",
              "unauthorized": "25"
            },
            "ledger_sequence": 58000000,
            "ledger_closed_at": "2026-08-10T12:00:00.000Z"
          },
          "confidence": 0.95,
          "confidence_formula_version": "trustline-state-confidence-v0.1",
          "confidence_components": {
            "agreement": 1,
            "freshness": 0.95,
            "availability": 1,
            "spread": 1
          },
          "confidence_caps_applied": [],
          "sources_configured": 2,
          "sources_responded": 2,
          "sources_usable": 2,
          "sources_agreeing": 2,
          "sources_excluded": 0,
          "contributions": [],
          "discrepancies": [],
          "source_errors": [],
          "as_of": "2026-08-10T12:00:00.000Z",
          "methodology_version": "trustline-state-v0.1",
          "request_id": "example_request",
          "api_version": "v1"
        }
      },
      "TrustlineDegraded": {
        "summary": "Degraded trustline state",
        "value": {
          "metric": "trustline_state",
          "subject": {
            "kind": "asset",
            "asset": "USDC:GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
          },
          "status": "degraded",
          "value": {
            "kind": "trustline_state",
            "total": "825",
            "states": {
              "authorized": "700",
              "authorized_to_maintain_liabilities": "100",
              "unauthorized": "25"
            },
            "ledger_sequence": 58000000,
            "ledger_closed_at": "2026-08-10T12:00:00.000Z"
          },
          "confidence": 0.6,
          "confidence_formula_version": "trustline-state-confidence-v0.1",
          "confidence_components": {
            "agreement": 1,
            "freshness": 0.95,
            "availability": 0.5,
            "spread": 1
          },
          "confidence_caps_applied": [
            "single_source"
          ],
          "sources_configured": 2,
          "sources_responded": 1,
          "sources_usable": 1,
          "sources_agreeing": 1,
          "sources_excluded": 0,
          "contributions": [],
          "discrepancies": [],
          "source_errors": [],
          "as_of": "2026-08-10T12:00:00.000Z",
          "methodology_version": "trustline-state-v0.1",
          "request_id": "example_request",
          "api_version": "v1"
        }
      },
      "TrustlineUnavailable": {
        "summary": "Unavailable or stale trustline state",
        "value": {
          "metric": "trustline_state",
          "subject": {
            "kind": "asset",
            "asset": "USDC:GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
          },
          "status": "unavailable",
          "value": null,
          "confidence": 0,
          "confidence_formula_version": "trustline-state-confidence-v0.1",
          "confidence_components": {
            "agreement": 0,
            "freshness": 0,
            "availability": 0,
            "spread": 0
          },
          "confidence_caps_applied": [],
          "sources_configured": 2,
          "sources_responded": 0,
          "sources_usable": 0,
          "sources_agreeing": 0,
          "sources_excluded": 0,
          "contributions": [],
          "discrepancies": [],
          "source_errors": [
            {
              "source_id": null,
              "source_url": null,
              "code": "stale_observation",
              "category": "freshness",
              "message": "Latest finalized trustline evidence exceeds 900 seconds",
              "occurred_at": "2026-08-10T12:00:00.000Z",
              "retryable": false
            }
          ],
          "as_of": "2026-08-10T12:00:00.000Z",
          "methodology_version": "trustline-state-v0.1",
          "request_id": "example_request",
          "api_version": "v1"
        }
      },
...
      "InvalidRequestId": {
        "summary": "Invalid request identifier",
        "value": {
          "error": {
            "code": "invalid_request_id",
            "message": "X-Request-ID must be a valid identifier of at most 128 characters"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
      "InvalidQueryParameter": {
        "summary": "Unsupported query parameter",
        "value": {
          "error": {
            "code": "invalid_query_parameter",
            "message": "Unsupported query parameter: limit"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
      "InvalidAsset": {
        "summary": "Invalid supply asset",
        "value": {
          "error": {
            "code": "invalid_asset",
            "message": "Asset must be a canonical CODE:ISSUER credit-asset identifier"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
...
      "TrustlineMissingSnapshot": {
        "summary": "No finalized trustline snapshot",
        "value": {
          "error": {
            "code": "trustline_snapshot_not_found",
            "message": "No finalized trustline snapshot is available"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
...
      "TrustlineReadUnavailable": {
        "summary": "Trustline read storage unavailable",
        "value": {
          "error": {
            "code": "trustline_read_unavailable",
            "message": "The trustline read model is temporarily unavailable"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
...
      "AuthenticationError": {
        "summary": "Authentication is required but the supplied API key is not usable",
        "value": {
          "error": {
            "code": "authentication_required",
            "message": "A valid API key is required"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
      "RateLimitError": {
        "summary": "The current API-key plan window is exhausted",
        "value": {
          "error": {
            "code": "rate_limit_exceeded",
            "message": "The request quota has been exceeded"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Axiom-Key",
        "description": "Required by the hosted API contract. Plaintext keys are shown only when issued."
      }
    }
  }
}